fix: rename and delete files.
